requirements.txt
black==25.12.0
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
idna==3.11
kaggle==1.8.3
kagglesdk==0.1.14
librt==0.7.5
mypy==1.19.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
protobuf==6.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytokens==0.3.0
requests==2.32.5
six==1.17.0
text-unidecode==1.3
tqdm==4.67.1
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing_extensions==4.15.0
urllib3==2.6.2
webencodings==0.5.1
